refactor: move more logic into `BulkLoadPayload`
